Updating ScrollListDirector to follow the "Avoiding huge director syndrome" design pattern - which is, to factor out a given show functionality into a different class in which the show can access via Director.
This example is not using NamedCommand for setup and exit of a particular show segment, but it is encapsulating java side's scroll list functioality into a class ScrollList.java instead of having them in the Director itself.  The scrolllist show file, specifically various RCHanddler commands for scrolling,  access ScrollList via
   java_command [[ getDirector().scrollList.XXX(); ]]

where XXX() is a method name in the ScrollList.java.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@604 e5771d8f-0841-e040-ffdb-cc069419ffcf