Tweaked grinview so that it will show a background image while show is running.
Frankly, I forget why this was done the other way.  I think it was perhaps
because the draw performance is quite bad when a background image is
showing.  However, if the person running grinview suffers from this, they
can just go to the UI and set the background back to blank anyway!  

At the moment, I'm working on a project that puts black images over a white
background, and that doesn't work so well with GrinView's default black 
background :-)



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@619 e5771d8f-0841-e040-ffdb-cc069419ffcf