Some miscellaneous cleanups:
Fixed a bug in FontStripText.destroy() when loading fails, that is,
when it doesn't find the font strip info file.  Also made GrinView
a little more robust when, say, there's a NullPointerException in
a place like FontStripText.destroy() :-)

Changed the GrinView jar build so it includes all .txt files in the source
directory, which help a bit when you $include.

Fixed a documentation typo in GrinXlet




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@623 e5771d8f-0841-e040-ffdb-cc069419ffcf