Whitespace cleanup.  Expanded tabs in source files, and converted all end-of-line delimiters to Unix.
The cookbook source used the JavaSoft coding convention, of an indent
level of four spaces, and tabs that expand to eight.  Many other projects
follow the default eclipse setting of tabs expanding to four spaces...
Sadly, eclipse followed the Windows/IDE convention, and it's caught on.

These days, The Right Thing to do is to always expand tabs.  In eclipse,
that can be set in the editor through preferences -> genera -> editors ->
text editors, then click on "insert spaces for tabs."  In vim, add 
"set expand" to your .vimrc, or (if you're like me) your .exrc.

While I was modifying every source file, I also changed all of the line
delimiters to the Unix convention of \n.  Some of the cookbook files were
originally created on Windows machines, so they got \r\n.

I converted this all with the following shell commands:

    for f in `find . -name '*.java' -print` `find . -name '*.txt' -print` ; do 
        expand $f | tr -d '\r' > ~/tmp/out.txt
        mv ~/tmp/out.txt $f
        echo $f
    done






git-svn-id: https://svn.java.net/svn/hdcookbook~svn@643 e5771d8f-0841-e040-ffdb-cc069419ffcf