Add support for KEY_TYPED event delivery to Directors.  No known Blu-ray players generate KEY_TYPED at present, but adding this now adds some future-proofing, and might potentially be useful on some other platforms (e.g. Android?).  The overhead for adding this support is a trivial amount of code, and no new classes.
Change GrinXlet so that it responds to mouse pressed events, rather than mouse
clicked events.  This makes it so the sequence click-move-release is
recognized as a press.  That sequence is a "drag," not a "click," so what
a button really wants to know about is mouse presses, not mouse clicks.

Changed internal naming thoughout to "press" rather than "click."

Added a build property to GrinXlet to turn on the mouse event reception
flags in the BDJO.

Reviewed by Chihiro and Joe.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@652 e5771d8f-0841-e040-ffdb-cc069419ffcf