Make support for mouse events consistent with key pressed events.  Now, a Show consumes a mouse event that it processes via the RC handler mechanism, as is done with key pressed events.
This change might introduce a slight incompatibility in the area of mouse
handling with GRIN extensions.  We added an abstract method to RCHandler,
and the Show.handleMouseXXX() methods now return a boolean and take another
argument.  Also, the Show now handles the re-dispatch of mouse events to the 
animation thread, rather than this being done at the GrinXlet level.

Reviewed by Chihiro.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@655 e5771d8f-0841-e040-ffdb-cc069419ffcf